+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Version 4:
 - Optimize automatic refresh to use partial updates instead of fully rescanning the document. This improves performance significantly on large documents.
 - Add option to toggle automatic refresh on/off, Add button to manually refresh the view
 - Add possibility to cycle through instances of a component variant
+- Restyle Component variants view to be more table-like for better readability